+++ lib/dart_failure_handler.dart
@@ -2,5 +2,5 @@ library;
 
 export 'src/either.dart';
 export 'src/failure.dart';
-export 'src/error_handler.dart';
+export 'src/error_mapper.dart';
 export 'src/repository_mixin.dart';